Added behavior for removing upgrade chips. Add hotbar display for heat data component
Took 52 minutes

+++ src/main/java/com/github/x3r/mekanism_weaponry/common/registry/DataComponentRegistry.java
@@ -36,9 +36,9 @@ public class DataComponentRegistry {
             builder -> builder.persistent(Codec.INT).networkSynchronized(ByteBufCodecs.INT)
     );
 
-    public static final DeferredHolder<DataComponentType<?>, DataComponentType<Integer>> HEAT = DATA_COMPONENTS.registerComponentType(
+    public static final DeferredHolder<DataComponentType<?>, DataComponentType<Float>> HEAT = DATA_COMPONENTS.registerComponentType(
             "heat",
-            builder -> builder.persistent(Codec.INT).networkSynchronized(ByteBufCodecs.INT)
+            builder -> builder.persistent(Codec.FLOAT).networkSynchronized(ByteBufCodecs.FLOAT)
     );
 
 }
